Improve demo INP and docs SEO without package changes.
Code-split kit adapters behind startTransition, remount only on kit switch,
and surface install/StackBlitz CTAs at the switcher. Rewrite SERP titles/metas
across docs and tighten landing/demo paint and conversion paths.
