+++ apps/showcase/columns/index.html
@@ -6,8 +6,13 @@
     <link rel="icon" type="image/svg+xml" href="../favicon.svg" />
     <meta
       name="description"
-      content="Show, hide, drag-reorder, pin left or right, and resize columns by drag or keyboard — then persist the layout to the URL or storage."
+      content="AdaptTable column management demo — show/hide, drag-reorder, pin, and resize columns, then persist the layout to the URL."
     />
+    <link
+      rel="canonical"
+      href="https://orwa-mahmoud.github.io/adapttable/demo/columns/"
+    />
+    <meta name="robots" content="index, follow, max-image-preview:large" />
     <!-- Cloudflare Web Analytics — cookieless, no consent banner needed. -->
     <script
       defer
@@ -17,13 +22,17 @@
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
-      href="https://fonts.googleapis.com/css2?family=Schibsted+Grotesk:wght@400;500;600;700;800&family=JetBrains+Mono:wght@400;500;600&family=IBM+Plex+Sans+Arabic:wght@400;600;700&display=swap"
+      href="https://fonts.googleapis.com/css2?family=Schibsted+Grotesk:wght@400;500;600;700;800&display=swap"
       rel="stylesheet"
     />
-    <title>AdaptTable demo — column management</title>
+    <link
+      href="https://fonts.googleapis.com/css2?family=JetBrains+Mono:wght@400;500;600&family=IBM+Plex+Sans+Arabic:wght@400;600;700&display=swap"
+      rel="stylesheet"
+      media="print"
+      onload="this.media = 'all'"
+    />
+    <title>AdaptTable demo — column pin, resize, reorder</title>
     <script>
-      // Apply the stored theme before paint so page-to-page navigation
-      // doesn't flash light.
       try {
         if (localStorage.getItem("adapttable-demo-theme") === "dark")
           document.documentElement.dataset.theme = "dark";
