+++ apps/showcase/src/main.tsx
@@ -5,7 +5,8 @@ import "./styles.css";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import { LiveDemo, PageShell } from "./App";
+import { LiveDemo } from "./LiveDemo";
+import { PageShell } from "./PageShell";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
