+++ docs/sorting.md
@@ -1,4 +1,4 @@
-# Sorting
+# React table sorting — single, multi-column, server-side
 
 ▶ **Try it live:** [open a Mantine starter in StackBlitz](https://stackblitz.com/github/orwa-mahmoud/adapttable/tree/main/starters/mantine) — a real AdaptTable you can edit in the browser, no install. [Other UI kits →](./getting-started.md#try-it-in-stackblitz)
